refactor: Refactoring all entity
