+++ src/main/java/com/deu/synabro/repository/EducationRepository.java
@@ -1,5 +1,6 @@
-package com.deu.synabro.education;
+package com.deu.synabro.repository;
 
+import com.deu.synabro.entity.Education;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
