+++ src/main/java/com/deu/synabro/repository/InspectionRepository.java
@@ -1,8 +1,9 @@
-package com.deu.synabro.volunteer;
+package com.deu.synabro.repository;
 
+import com.deu.synabro.entity.Inspection;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface VolunteerRepository extends JpaRepository<Volunteer, Long> {
+public interface InspectionRepository extends JpaRepository<Inspection, Long> {
 }
