+++ src/main/java/com/deu/synabro/repository/VolunteerRepository.java
@@ -1,8 +1,9 @@
-package com.deu.synabro.inspection;
+package com.deu.synabro.repository;
 
+import com.deu.synabro.entity.Volunteer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface InspectionRepository extends JpaRepository<Inspection, Long> {
+public interface VolunteerRepository extends JpaRepository<Volunteer, Long> {
 }
